Replaced TimerB UserInputs with Input Interrupts to save power. Utilised TimerA to check for continuous holding down on inputs.
Deleted TimerB functionality, and replaced with Interrupt for User
Inputs. In this way user inputs will be serviced when they happen,
but if user does not input anything, power will be saved as TimerB
will not be periodically checking anymore.

This created the issue of needing to check for continuous holding down
of user input, which was taken care of before by TimerB periodicity, so
created a function to check for continuous pressing.