Added 2nd player, scores screen, ability to continue after scoring.
Added 2nd player on the right of the LCD, able to be controlled via
the SW1 and SW1 buttons for UP and DOWN respectively, made rackets
1 pixel less wide to help maintain speed.
Added score screen which displays goal scorer and current scores
when a goal is scored. It also displays 'Reset to start' which was
already there, but also now 'Press an input to continue' which lets
the game continue by pressing any of JOYSTICK UP/DOWN or Buttons
UP/DOWN. The direction of the ball once continuing will travel away
from the last goalscorer.