Changed movement rules depending on if racket is moving or not.
If racket is moving then the ball is affected by the movement direction
of the racket.
If rack is not moving then the ball is affected by which part of the
racket it strikes. Centre part will keep current y_displacement of the ball;
lower part will increase ball y_displacement in downward direction;
upper part of the racket will increase ball y_displacement in an upward
direction. This means that y_displacements are able to be cancelled out,
kept the same, or increased depending on the player's choice to affect
the ball movement with either racket movement or with stationary
directional racket zones.