Added End-of-Game functionality. Reach winning scoreline to win game.
Winning scoreline is hard-coded into the game, but the first player
to reach it wins the game. A splash screen showing the winner shows.
A message tells the user(s) that by pressing an input or the reset
button, they can continue, this will restart the game and show the
intro menu for a new game. Tested and working.