Created shooter bonus for Player1
Once either p1 or p2 reaches a score of 1, a bonus appears which
when hit with the ball gives that hitter a bonus. In the case of
player1 that is cannons at both ends of the racket which are
triggered by pressing JOYSTICK RIGHT. On triggering the twin
cannons will fire quantum-linked plasma rounds, and if they strike
the ball they cause it to flip it's x and y direction. This can
be beneficial for getting the ball past your opponent, however it
is possible accidentally flip the ball toward the wall you are
defending too.