Scoring updates no longer interrupt game. Winning does.
Score updates have been changed so instead of showing a
splash screen, instead scores are updated as goals happen,
with a scorecard kept in a title banner at the top of the
screen. This meant adjusted the racket and ball limits to
not encroach on the top banner. It also required resetting
racket positions when a goal is scored, and the coolest way
seemed to be to move them back into centre position before
allowing normal play to continue, so they will slide back
into position with the ball waiting in the centre, when ready
the ball will move again and players will regain control of
their rackets again. The winning splash screen still works.