Improved on-screen notifications for scoring, bonuses, movement
Edited banner to display p1 and p2 to make scores more intuitive
to understand. Edited banner to include 'B' next to a player if
they have a bonus currently enabled, e.g. if they picked up a bonus.
Bonus 'B' is also on a cooldown which decrements each time the
opponent scores, so the bonus deactives after the opponent scores three
times. Including a 'GOAL SCORED' message in the middle of the LCD to
show while the rackets are moving back to their original positions, it
feels better to have something showing there while the game is paused,
to explain the pause and keep the players involved and informed.