Created player2 bonus, a bonus extra mini racket.
The bonus racket moves in the opposite direction to the main racket.
It has racket movement effect, e.g. if ball strikes will racket is
moving the ball will gain y_displacement in the same direction.
It lacks directional zones, so if it stationary the ball will simply
reflect off it.